Мелкие правки.

+++ Configs/pgsql.leoka_estetica_dev/Tables/[Configs].[ProjectVacancyColumnsNames]/[Configs].[ProjectVacancyColumnsNames]_CreateTable.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "Configs"."ProjectVacancyColumnsNames"
 (
-    "ColumnId" BIGINT,
+    "ColumnId" SERIAL,
     "ColumnName" VARCHAR(200) NOT NULL,
     "TableName" VARCHAR(200) NOT NULL,
     "Position" INT NOT NULL,
